% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_mars.R
\name{basis_table_mars}
\alias{basis_table_mars}
\title{Termes (fonctions charnière) du modèle final et leurs coefficients}
\usage{
basis_table_mars(fit)
}
\arguments{
\item{fit}{objet "mars_cv_fit"}
}
\description{
Coefficient exprimé dans l'unité de SA variable (non comparable entre
variables) ; log-odds en binomial.
}
