% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{build_reduced_formula}
\alias{build_reduced_formula}
\title{Construit une formule "réponse ~ variables" sûre (backticks partout)}
\usage{
build_reduced_formula(variables, response)
}
\arguments{
\item{variables}{vecteur de noms (ex. \code{coef_table_lasso(fit)$VariableOriginale})}

\item{response}{nom de la réponse}
}
\value{
formule
}
\description{
Construit une formule "réponse ~ variables" sûre (backticks partout)
}
