% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_mars.R
\name{check_complexity_mars}
\alias{check_complexity_mars}
\title{La non-linéarité est-elle réellement utilisée par MARS ?}
\usage{
check_complexity_mars(fit)
}
\arguments{
\item{fit}{objet "mars_cv_fit"}
}
\description{
Part des folds avec au moins une charnière / une interaction. Presque
aucune charnière = MARS se comporte en modèle linéaire creux.
}
