% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_residuals.R
\name{check_heteroscedasticity}
\alias{check_heteroscedasticity}
\title{Homogénéité de la variance des résidus (Spearman |résidu| vs prédiction)}
\usage{
check_heteroscedasticity(fit)
}
\arguments{
\item{fit}{objet "cv_fit" (family = "gaussian")}
}
\value{
liste de classe "check_heteroscedasticity_cv"
}
\description{
Homogénéité de la variance des résidus (Spearman |résidu| vs prédiction)
}
