% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{check_hyperparam_stability}
\alias{check_hyperparam_stability}
\title{Stabilité des hyperparamètres à travers les folds}
\usage{
check_hyperparam_stability(fit)
}
\arguments{
\item{fit}{objet "cv_fit"}
}
\description{
Stabilité des hyperparamètres à travers les folds
}
