% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_residuals.R
\name{check_independence}
\alias{check_independence}
\title{Cohérence du résidu d'un même individu entre répétitions}
\usage{
check_independence(fit)
}
\arguments{
\item{fit}{objet "cv_fit"}
}
\value{
liste de classe "check_independence_cv"
}
\description{
Cohérence du résidu d'un même individu entre répétitions
}
