% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_lasso.R
\name{check_lambda_range_lasso}
\alias{check_lambda_range_lasso}
\alias{plot_lambda_path_lasso}
\alias{plot_lambda_stability}
\title{Lambda.min au bord de sa grille ? (modèle final et folds)}
\usage{
check_lambda_range_lasso(fit)

plot_lambda_path_lasso(fit)

plot_lambda_stability(fit)
}
\arguments{
\item{fit}{objet "lasso_cv_fit"}
}
\description{
Lambda.min au bord de sa grille ? (modèle final et folds)
}
