% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_mars.R
\name{check_selection_mars}
\alias{check_selection_mars}
\title{Stabilité de la sélection des variables par MARS}
\usage{
check_selection_mars(fit, top_n = 30)
}
\arguments{
\item{fit}{objet "mars_cv_fit"}

\item{top_n}{variables conservées (30)}
}
\description{
Stabilité de la sélection des variables par MARS
}
