% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_lasso.R
\name{coef_table_lasso}
\alias{coef_table_lasso}
\title{Coefficients non nuls du modèle final (jamais tronqués)}
\usage{
coef_table_lasso(fit)
}
\arguments{
\item{fit}{objet "lasso_cv_fit"}
}
\value{
tibble Variable / Coefficient / VariableOriginale, par |coefficient| décroissant
}
\description{
Coefficients non nuls du modèle final (jamais tronqués)
}
