% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_calibration}
\alias{compare_calibration}
\title{Superpose les courbes de calibration de plusieurs modèles binomiaux}
\usage{
compare_calibration(fits, labels = NULL, bins = 10)
}
\arguments{
\item{fits}{liste de fits binomiaux}

\item{labels}{libellés optionnels}

\item{bins}{nombre de tranches (10)}
}
\description{
Superpose les courbes de calibration de plusieurs modèles binomiaux
}
