% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_computational_cost}
\alias{compare_computational_cost}
\title{Temps de calcul par modèle (+ SHAP optionnel)}
\usage{
compare_computational_cost(fits, labels = NULL, shap_result = NULL)
}
\arguments{
\item{fits}{liste d'au moins 1 fit}

\item{labels}{libellés optionnels}

\item{shap_result}{sortie de compare_shap() (ou check_shap() si un seul fit)}
}
\value{
liste de classe "compare_computational_cost" (table, plot)
}
\description{
Temps de calcul par modèle (+ SHAP optionnel)
}
