% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_roc}
\alias{compare_roc}
\title{Superpose les courbes ROC de plusieurs modèles binomiaux}
\usage{
compare_roc(fits, labels = NULL)
}
\arguments{
\item{fits}{liste de fits binomiaux}

\item{labels}{libellés optionnels}
}
\description{
Superpose les courbes ROC de plusieurs modèles binomiaux
}
