% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{cvperf-compat}
\alias{cvperf-compat}
\alias{check_normality_lasso}
\alias{check_heteroscedasticity_lasso}
\alias{check_outliers_lasso}
\alias{check_independence_lasso}
\alias{plot_roc_lasso}
\alias{plot_precision_recall_lasso}
\alias{plot_calibration_lasso}
\alias{plot_confusion_matrix_lasso}
\alias{check_model_lasso}
\alias{plot.check_normality_lasso}
\alias{plot.check_heteroscedasticity_lasso}
\alias{plot.check_outliers_lasso}
\alias{plot.check_independence_lasso}
\alias{compare_models_lasso}
\alias{check_normality_rf}
\alias{check_heteroscedasticity_rf}
\alias{check_outliers_rf}
\alias{check_independence_rf}
\alias{plot_roc_rf}
\alias{plot_precision_recall_rf}
\alias{plot_calibration_rf}
\alias{plot_confusion_matrix_rf}
\alias{check_model_rf}
\alias{plot_prediction_quality_rf}
\alias{plot_residuals_vs_fitted_rf}
\alias{plot_residuals_distribution_rf}
\alias{plot.check_normality_rf}
\alias{plot.check_heteroscedasticity_rf}
\alias{plot.check_outliers_rf}
\alias{plot.check_independence_rf}
\alias{check_hyperparam_stability_rf}
\alias{plot_hyperparam_path_rf}
\alias{plot_hyperparam_stability_rf}
\alias{plot_variable_importance_rf}
\alias{plot_effects_rf}
\alias{importance_table_rf}
\alias{check_importance_rf}
\alias{plot_importance_rf}
\alias{compare_models_rf}
\alias{check_normality_xgb}
\alias{check_heteroscedasticity_xgb}
\alias{check_outliers_xgb}
\alias{check_independence_xgb}
\alias{plot_roc_xgb}
\alias{plot_precision_recall_xgb}
\alias{plot_calibration_xgb}
\alias{plot_confusion_matrix_xgb}
\alias{check_model_xgb}
\alias{plot_prediction_quality_xgb}
\alias{plot_residuals_vs_fitted_xgb}
\alias{plot_residuals_distribution_xgb}
\alias{plot.check_normality_xgb}
\alias{plot.check_heteroscedasticity_xgb}
\alias{plot.check_outliers_xgb}
\alias{plot.check_independence_xgb}
\alias{check_hyperparam_stability_xgb}
\alias{plot_hyperparam_path_xgb}
\alias{plot_hyperparam_stability_xgb}
\alias{plot_variable_importance_xgb}
\alias{plot_effects_xgb}
\alias{importance_table_xgb}
\alias{check_importance_xgb}
\alias{plot_importance_xgb}
\alias{group_importance_xgb}
\alias{check_normality_svm}
\alias{check_heteroscedasticity_svm}
\alias{check_outliers_svm}
\alias{check_independence_svm}
\alias{plot_roc_svm}
\alias{plot_precision_recall_svm}
\alias{plot_calibration_svm}
\alias{plot_confusion_matrix_svm}
\alias{check_model_svm}
\alias{plot_prediction_quality_svm}
\alias{plot_residuals_vs_fitted_svm}
\alias{plot_residuals_distribution_svm}
\alias{plot.check_normality_svm}
\alias{plot.check_heteroscedasticity_svm}
\alias{plot.check_outliers_svm}
\alias{plot.check_independence_svm}
\alias{check_hyperparam_stability_svm}
\alias{plot_hyperparam_path_svm}
\alias{plot_hyperparam_stability_svm}
\alias{plot_variable_importance_svm}
\alias{plot_effects_svm}
\alias{importance_table_svm}
\alias{check_importance_svm}
\alias{plot_importance_svm}
\alias{group_importance_svm}
\alias{check_normality_mars}
\alias{check_heteroscedasticity_mars}
\alias{check_outliers_mars}
\alias{check_independence_mars}
\alias{plot_roc_mars}
\alias{plot_precision_recall_mars}
\alias{plot_calibration_mars}
\alias{plot_confusion_matrix_mars}
\alias{check_model_mars}
\alias{plot_prediction_quality_mars}
\alias{plot_residuals_vs_fitted_mars}
\alias{plot_residuals_distribution_mars}
\alias{plot.check_normality_mars}
\alias{plot.check_heteroscedasticity_mars}
\alias{plot.check_outliers_mars}
\alias{plot.check_independence_mars}
\alias{check_hyperparam_stability_mars}
\alias{plot_hyperparam_path_mars}
\alias{plot_hyperparam_stability_mars}
\alias{plot_variable_importance_mars}
\alias{plot_effects_mars}
\alias{importance_table_mars}
\alias{check_hyperparam_stability_lasso}
\title{Anciens noms des toolkits (compatibilite)}
\usage{
check_normality_lasso(fit, ...)

check_heteroscedasticity_lasso(fit, ...)

check_outliers_lasso(fit, ...)

check_independence_lasso(fit, ...)

plot_roc_lasso(fit, ...)

plot_precision_recall_lasso(fit, ...)

plot_calibration_lasso(fit, ...)

plot_confusion_matrix_lasso(fit, ...)

check_model_lasso(fit, ...)

\method{plot}{check_normality_lasso}(x, ...)

\method{plot}{check_heteroscedasticity_lasso}(x, ...)

\method{plot}{check_outliers_lasso}(x, ...)

\method{plot}{check_independence_lasso}(x, ...)

compare_models_lasso(fits, ...)

check_normality_rf(fit, ...)

check_heteroscedasticity_rf(fit, ...)

check_outliers_rf(fit, ...)

check_independence_rf(fit, ...)

plot_roc_rf(fit, ...)

plot_precision_recall_rf(fit, ...)

plot_calibration_rf(fit, ...)

plot_confusion_matrix_rf(fit, ...)

check_model_rf(fit, ...)

plot_prediction_quality_rf(fit, ...)

plot_residuals_vs_fitted_rf(fit, ...)

plot_residuals_distribution_rf(fit, ...)

\method{plot}{check_normality_rf}(x, ...)

\method{plot}{check_heteroscedasticity_rf}(x, ...)

\method{plot}{check_outliers_rf}(x, ...)

\method{plot}{check_independence_rf}(x, ...)

check_hyperparam_stability_rf(fit)

plot_hyperparam_path_rf(fit)

plot_hyperparam_stability_rf(fit)

plot_variable_importance_rf(fit, ...)

plot_effects_rf(fit, ...)

importance_table_rf(fit, ...)

check_importance_rf(fit, ...)

plot_importance_rf(x, ...)

compare_models_rf(fits, ...)

check_normality_xgb(fit, ...)

check_heteroscedasticity_xgb(fit, ...)

check_outliers_xgb(fit, ...)

check_independence_xgb(fit, ...)

plot_roc_xgb(fit, ...)

plot_precision_recall_xgb(fit, ...)

plot_calibration_xgb(fit, ...)

plot_confusion_matrix_xgb(fit, ...)

check_model_xgb(fit, ...)

plot_prediction_quality_xgb(fit, ...)

plot_residuals_vs_fitted_xgb(fit, ...)

plot_residuals_distribution_xgb(fit, ...)

\method{plot}{check_normality_xgb}(x, ...)

\method{plot}{check_heteroscedasticity_xgb}(x, ...)

\method{plot}{check_outliers_xgb}(x, ...)

\method{plot}{check_independence_xgb}(x, ...)

check_hyperparam_stability_xgb(fit)

plot_hyperparam_path_xgb(fit)

plot_hyperparam_stability_xgb(fit)

plot_variable_importance_xgb(fit, ...)

plot_effects_xgb(fit, ...)

importance_table_xgb(fit, ...)

check_importance_xgb(fit, ...)

plot_importance_xgb(x, ...)

group_importance_xgb(fit)

check_normality_svm(fit, ...)

check_heteroscedasticity_svm(fit, ...)

check_outliers_svm(fit, ...)

check_independence_svm(fit, ...)

plot_roc_svm(fit, ...)

plot_precision_recall_svm(fit, ...)

plot_calibration_svm(fit, ...)

plot_confusion_matrix_svm(fit, ...)

check_model_svm(fit, ...)

plot_prediction_quality_svm(fit, ...)

plot_residuals_vs_fitted_svm(fit, ...)

plot_residuals_distribution_svm(fit, ...)

\method{plot}{check_normality_svm}(x, ...)

\method{plot}{check_heteroscedasticity_svm}(x, ...)

\method{plot}{check_outliers_svm}(x, ...)

\method{plot}{check_independence_svm}(x, ...)

check_hyperparam_stability_svm(fit)

plot_hyperparam_path_svm(fit)

plot_hyperparam_stability_svm(fit)

plot_variable_importance_svm(fit, ...)

plot_effects_svm(fit, ...)

importance_table_svm(fit, ...)

check_importance_svm(fit, ...)

plot_importance_svm(x, ...)

group_importance_svm(fit)

check_normality_mars(fit, ...)

check_heteroscedasticity_mars(fit, ...)

check_outliers_mars(fit, ...)

check_independence_mars(fit, ...)

plot_roc_mars(fit, ...)

plot_precision_recall_mars(fit, ...)

plot_calibration_mars(fit, ...)

plot_confusion_matrix_mars(fit, ...)

check_model_mars(fit, ...)

plot_prediction_quality_mars(fit, ...)

plot_residuals_vs_fitted_mars(fit, ...)

plot_residuals_distribution_mars(fit, ...)

\method{plot}{check_normality_mars}(x, ...)

\method{plot}{check_heteroscedasticity_mars}(x, ...)

\method{plot}{check_outliers_mars}(x, ...)

\method{plot}{check_independence_mars}(x, ...)

check_hyperparam_stability_mars(fit)

plot_hyperparam_path_mars(fit)

plot_hyperparam_stability_mars(fit)

plot_variable_importance_mars(fit, ...)

plot_effects_mars(fit, ...)

importance_table_mars(fit, ...)

check_hyperparam_stability_lasso(fit)
}
\arguments{
\item{fit, x, fits}{objets d'entree}

\item{...}{arguments transmis}
}
\description{
Alias vers les fonctions communes, pour que les scripts ecrits avec les
toolkits \code{*_performance_toolkit.R} fonctionnent sans modification.
}
