% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{method_notes}
\alias{method_notes}
\title{Avertissements d'interprétation propres à une méthode}
\usage{
method_notes(fit, topic)
}
\arguments{
\item{fit}{objet "cv_fit"}

\item{topic}{"normality", "heteroscedasticity", "calibration", "final",
"residuals"}
}
\value{
chaîne de caractères ou NULL
}
\description{
Avertissements d'interprétation propres à une méthode
}
