% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_to_stars}
\alias{p_to_stars}
\title{Étoiles de significativité conventionnelles}
\usage{
p_to_stars(p)
}
\arguments{
\item{p}{vecteur de p-values}
}
\value{
vecteur character ("***", "**", "*", ".", "ns")
}
\description{
Étoiles de significativité conventionnelles
}
