% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_lasso.R
\name{plot_collinearity}
\alias{plot_collinearity}
\title{Lollipop du VIF (vert < 5, bleu 5-10, rouge >= 10)}
\usage{
plot_collinearity(x, labels = NULL, log_scale = TRUE)
}
\arguments{
\item{x}{objet "check_collinearity_lasso"}

\item{labels}{libellés optionnels}

\item{log_scale}{axe log10 (TRUE)}
}
\description{
Lollipop du VIF (vert < 5, bleu 5-10, rouge >= 10)
}
