% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_classification.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Matrice de confusion (prédictions poolées)}
\usage{
plot_confusion_matrix(fit, threshold = NULL)
}
\arguments{
\item{fit}{objet "cv_fit" (family = "binomial")}

\item{threshold}{seuil ; NULL = \code{fit$classification_threshold}}
}
\description{
Matrice de confusion (prédictions poolées)
}
