% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_hyperparam_path}
\alias{plot_hyperparam_path}
\title{Chemin de réglage des hyperparamètres (modèle final)}
\usage{
plot_hyperparam_path(fit)
}
\arguments{
\item{fit}{objet "cv_fit"}
}
\description{
Chemin de réglage des hyperparamètres (modèle final)
}
