% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_hyperparam_stability}
\alias{plot_hyperparam_stability}
\title{Distribution des hyperparamètres retenus à travers les folds}
\usage{
plot_hyperparam_stability(fit)
}
\arguments{
\item{fit}{objet "cv_fit"}
}
\description{
Distribution des hyperparamètres retenus à travers les folds
}
