% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_classification.R
\name{plot_precision_recall}
\alias{plot_precision_recall}
\title{Courbe précision-rappel (prédictions poolées)}
\usage{
plot_precision_recall(fit)
}
\arguments{
\item{fit}{objet "cv_fit" (family = "binomial")}
}
\description{
Courbe précision-rappel (prédictions poolées)
}
