% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_prediction.R
\name{plot_residuals_distribution}
\alias{plot_residuals_distribution}
\title{Distribution des résidus poolés}
\usage{
plot_residuals_distribution(fit)
}
\arguments{
\item{fit}{objet "cv_fit" (family = "gaussian")}
}
\description{
Distribution des résidus poolés
}
