% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_prediction.R
\name{plot_residuals_vs_fitted}
\alias{plot_residuals_vs_fitted}
\title{Résidus vs valeurs prédites (résidus poolés de CV)}
\usage{
plot_residuals_vs_fitted(fit)
}
\arguments{
\item{fit}{objet "cv_fit" (family = "gaussian")}
}
\description{
Résidus vs valeurs prédites (résidus poolés de CV)
}
