% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_mars.R
\name{plot_selection_mars}
\alias{plot_selection_mars}
\title{Lollipop de la fréquence de sélection MARS}
\usage{
plot_selection_mars(x, labels = NULL)
}
\arguments{
\item{x}{objet "check_selection_mars"}

\item{labels}{libellés optionnels}
}
\description{
Lollipop de la fréquence de sélection MARS
}
