% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap.R
\name{plot_shap}
\alias{plot_shap}
\title{Représentation des valeurs de Shapley}
\usage{
plot_shap(x, labels = NULL, top_n = 20, type = c("beeswarm", "bar"))
}
\arguments{
\item{x}{objet "check_shap"}

\item{labels}{libellés optionnels}

\item{top_n}{variables affichées (défaut 20)}

\item{type}{"beeswarm" (1 point = 1 observation x 1 variable, couleur =
valeur du prédicteur) ou "bar" (|SHAP| moyen)}
}
\description{
Représentation des valeurs de Shapley
}
