% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{variables_summary}
\alias{variables_summary}
\title{Résumé "variables en entrée / en sortie" d'un fit (utilisé par compare_models())}
\usage{
variables_summary(fit)
}
\arguments{
\item{fit}{objet "cv_fit"}
}
\description{
Résumé "variables en entrée / en sortie" d'un fit (utilisé par compare_models())
}
